+++ ToolGood.ReadyGo3/PetaPoco/Providers/SqlServer2012DatabaseProvider.cs
@@ -9,7 +9,7 @@ namespace ToolGood.ReadyGo3.PetaPoco.Providers
 
         //private static readonly Regex SelectTopRegex = new Regex(@"^SELECT +TOP(\d+)", RegexOptions.IgnoreCase);
 
-        public override string BuildPageQuery(long skip, long take, SQLParts parts, ref object[] args)
+        public override string BuildPageQuery(int skip, int take, SQLParts parts, ref object[] args)
         {
             if (!parts.Sql.ToLower().Contains("order by")) {
                 return base.BuildPageQuery(skip, take, parts, ref args);
